fix: address feedback from greptile
