+++ extensions/gist-rocket/src/components/update-from-file.tsx
@@ -1,10 +1,8 @@
 import { Action, ActionPanel, Form, Icon, Toast, getPreferenceValues, showToast, useNavigation } from "@raycast/api";
 import { useState } from "react";
-import { readFile, stat } from "node:fs/promises";
-import { join } from "node:path";
 import { afterPublish, republish, PublishInput } from "../lib/publish";
-import { detectKind, inlineLocalAssets, suggestedTitle, wrapHtmlFragment, baseFileName } from "../lib/html";
-import { renderMarkdownToHtml, Theme } from "../lib/render";
+import { loadFromPath } from "../lib/file-loader";
+import { Theme } from "../lib/render";
 
 export function UpdateFromFile(props: {
   gistId: string;
@@ -71,47 +69,4 @@ export function UpdateFromFile(props: {
       </Form.Dropdown>
     </Form>
   );
-}
-
-async function loadFromPath(p: string, opts: { theme: Theme; description: string }): Promise<PublishInput> {
-  const s = await stat(p);
-  if (s.isDirectory()) {
-    const indexPath = join(p, "index.html");
-    const raw = await readFile(indexPath, "utf8").catch(() => {
-      throw new Error("Folder must contain an index.html");
-    });
-    const inlined = await inlineLocalAssets(raw, indexPath);
-    return {
-      kind: "html",
-      source: inlined,
-      preRenderedHtml: inlined,
-      description: opts.description || suggestedTitle(inlined, "Untitled page"),
-      visibility: "secret",
-      theme: opts.theme,
-    };
-  }
-  const raw = await readFile(p, "utf8");
-  const kind = detectKind(p, true);
-  if (kind === "markdown") {
-    const html = await renderMarkdownToHtml(raw, { theme: opts.theme, title: opts.description || undefined });
-    return {
-      kind: "markdown",
-      source: raw,
-      preRenderedHtml: html,
-      description: opts.description || suggestedTitle(raw, baseFileName(p)),
-      visibility: "secret",
-      theme: opts.theme,
-      sourceFilename: "source.md",
-    };
-  }
-  const inlined = await inlineLocalAssets(raw, p);
-  const wrapped = wrapHtmlFragment(inlined, opts.description || undefined);
-  return {
-    kind: "html",
-    source: wrapped,
-    preRenderedHtml: wrapped,
-    description: opts.description || suggestedTitle(wrapped, baseFileName(p)),
-    visibility: "secret",
-    theme: opts.theme,
-  };
 }
